+++ lib/src/models/overlay_manager_entry.dart
@@ -26,15 +26,24 @@ abstract class OverlayManagerEntry<T> extends Comparable {
   /// Used for internal! [isDismissible] is true if the overlay can be close automatically.
   final bool isDismissible;
 
-  /// In development, the [elevation] provides layer position of this overlay in manage
-  final double elevation;
+  /// The [elevation] provides the layer's z-index position of this overlay in manage.
+  /// Call [OverlayManager.rearrange] or use [OverlayManagerEntry.rearrange] to apply the new overlay elevation order.
+  ///
+  /// Note: Nothing happens if there is no the [rearrange] function is called.
+  double elevation;
+
+  /// Rearrange entry with [newElevation]
+  void rearrange(double newElevation) {
+    elevation = newElevation;
+    manager.rearrange();
+  }
 
   /// [closed] is true if this overlay is closed.
   /// each overlay can be shown only one time and can not be reused if it is closed
   bool get closed;
 
   /// close this overlay if [closed] is false and return T [value] for [onDismiss]
-  void close(T? value);
+  void close([T? value]);
 
   @override
   int compareTo(other) {