+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1
 
 initial
+
+## 0.0.2
+
+- Fix: OverlayManagerEntry generic type.
+- New Feature: Rearrange entry.